[#172989345] add actions add request handler
