+++ ts/features/bonusVacanze/store/sagas/startBonusEligibilitySaga.ts
@@ -1,0 +1,121 @@
+import { readableReport } from "italia-ts-commons/lib/reporters";
+import { Millisecond } from "italia-ts-commons/lib/units";
+import { SagaIterator } from "redux-saga";
+import { call, put } from "redux-saga/effects";
+import { SagaCallReturnType } from "../../../../types/utils";
+import { startTimer } from "../../../../utils/timer";
+import { BackendBonusVacanze } from "../../api/backendBonusVacanze";
+import {
+  checkBonusEligibility,
+  eligibilityRequestId,
+  eligibilityRequestProgress
+} from "../actions/bonusVacanze";
+import { EligibilityRequestProgressEnum } from "../reducers/bonusVacanze";
+
+const checkEligibilityResultPolling = 1000 as Millisecond;
+// stop polling when elapsed time from the beginning exceeds this threshold
+const pollingTimeThreshold = (10 * 1000) as Millisecond;
+
+// handle start bonus eligibility check
+function* checkBonusEligibilitySaga(
+  getEligibilityCheck: ReturnType<
+    typeof BackendBonusVacanze
+  >["getEligibilityCheck"]
+): SagaIterator {
+  try {
+    const eligibilityCheckResult: SagaCallReturnType<
+      typeof getEligibilityCheck
+    > = yield call(getEligibilityCheck, {});
+    if (eligibilityCheckResult.isRight()) {
+      // we got the check result
+      if (eligibilityCheckResult.value.status === 200) {
+        yield put(
+          checkBonusEligibility.success(eligibilityCheckResult.value.value)
+        );
+        return true;
+      }
+      return false;
+    } else {
+      // we got some error, stop polling
+      throw Error(readableReport(eligibilityCheckResult.value));
+    }
+  } catch (e) {
+    yield put(checkBonusEligibility.failure(e));
+    return false;
+  }
+}
+
+// handle start bonus eligibility check
+// tslint:disable-next-line: cognitive-complexity
+export function* startBonusEligibilitySaga(
+  postEligibilityCheck: ReturnType<
+    typeof BackendBonusVacanze
+  >["postEligibilityCheck"],
+  getEligibilityCheck: ReturnType<
+    typeof BackendBonusVacanze
+  >["getEligibilityCheck"]
+): SagaIterator {
+  try {
+    // request is pending
+    yield put(
+      eligibilityRequestProgress(EligibilityRequestProgressEnum.PENDING)
+    );
+    const startEligibilityResult: SagaCallReturnType<
+      typeof postEligibilityCheck
+    > = yield call(postEligibilityCheck, {});
+    if (startEligibilityResult.isRight()) {
+      // 202 -> request accepted | 409 -> pending request
+      if (
+        startEligibilityResult.value.status === 202 ||
+        startEligibilityResult.value.status === 409
+      ) {
+        // processing request, dispatch di process id
+        if (startEligibilityResult.value.status === 202) {
+          yield put(eligibilityRequestId(startEligibilityResult.value.value));
+        }
+        // start polling to know about the check result
+        const startPolling = new Date().getTime();
+        // request is pending
+        yield put(
+          eligibilityRequestProgress(
+            EligibilityRequestProgressEnum.POLLING_STARTED
+          )
+        );
+        // TODO: handle cancel request (stop polling)
+        while (true) {
+          const eligibilityCheckResult: boolean = yield call(
+            checkBonusEligibilitySaga,
+            getEligibilityCheck
+          );
+          // we got the response, stop polling
+          if (eligibilityCheckResult === true) {
+            yield put(
+              eligibilityRequestProgress(
+                EligibilityRequestProgressEnum.COMPLETE
+              )
+            );
+            return;
+          }
+          // sleep
+          yield call(startTimer, checkEligibilityResultPolling);
+          // check if the time threshold was exceeded, if yes abort
+          const now = new Date().getTime();
+          if (now - startPolling >= pollingTimeThreshold) {
+            yield put(
+              eligibilityRequestProgress(
+                EligibilityRequestProgressEnum.POLLING_EXCEEDED
+              )
+            );
+            return;
+          }
+        }
+      }
+      throw Error(`response status ${startEligibilityResult.value.status}`);
+    } else {
+      throw Error(readableReport(startEligibilityResult.value));
+    }
+  } catch (e) {
+    yield put(eligibilityRequestProgress(EligibilityRequestProgressEnum.ERROR));
+    yield put(checkBonusEligibility.failure(e));
+  }
+}
